Update environment variables syntax in .drone.yml
The commit is a minor update where we changed the way we reference the environment variables. Specifically, we replaced the "$VAR_NAME" syntax with "${VAR_NAME}" in the commands for the dotnet scanner in our .drone.yml file. This should make our code more consistent with common usage and avoids possible confusions or errors.